requirements.txt
PyQt5>=5.15.0
pyqtgraph>=0.12.0
pyserial>=3.5